Add UserConfigFile
UserConfigFile is an implementation of the abstract AdditionalText so that it is the same base type as the Analyzer's AdditionalFiles.

When the project is able to be upgrade to .NET Standard 2.1 UserConfigFile.FromApplicationData can be updated to use the async version of File.ReadAllText so that it can respect the analyzer context cancellation token.